+++ server/tests/api/api_stats_test.py
@@ -234,53 +234,3 @@ class TestSearchStatVar(unittest.TestCase):
             assert response.status_code == 200
             result = json.loads(response.data)
             assert result == expected_sv_only_result
-
-
-class TestGetVariableGroupInfo(unittest.TestCase):
-
-    @mock.patch('routes.api.variable_group.dc.get')
-    def test_statvar_path(self, mock_result):
-        expected_result = {
-            "absoluteName":
-                "Data Commons Variables",
-            "childStatVarGroups": [{
-                "id": "dc/g/Crime",
-                "specializedEntity": "Crime",
-                "displayName": "Crime"
-            }, {
-                "id": "dc/g/Demographics",
-                "specializedEntity": "Demographics",
-                "displayName": "Demographics"
-            }]
-        }
-
-        def side_effect(url):
-            if url == "/v1/info/variable-group/dc/g/Root?constrained_entities=geoId/06":
-                return {"info": expected_result}
-            else:
-                return {}
-
-        mock_result.side_effect = side_effect
-        response = app.test_client().get(
-            'api/variable-group/info?dcid=dc/g/Root&entities=geoId/06')
-        assert response.status_code == 200
-        result = json.loads(response.data)
-        assert result == expected_result
-
-
-class TestVariablePath(unittest.TestCase):
-
-    @mock.patch('routes.api.variable.dc.get')
-    def test_variable_path(self, mock_result):
-
-        def side_effect(url):
-            if url == "/v1/variable/ancestors/Count_Person":
-                return {"ancestors": ["dc/g/Demographics"]}
-            else:
-                return {}
-
-        mock_result.side_effect = side_effect
-        response = app.test_client().get('api/variable/path?dcid=Count_Person')
-        assert response.status_code == 200
-        result = json.loads(response.data)
-        assert result == ["Count_Person", "dc/g/Demographics"]
